move tools in dedicated razor class library
