+++ src/KITT.Web.App.Tools/Pages/Index.razor
@@ -4,7 +4,7 @@
 
 <PageLayout PageTitle="Tools">
     <Toolbar>
-        <NavButton Label="@CommonLocalizer[nameof(Resources.Common.Back)]"
+        <NavButton Label="@Localizer[nameof(Resources.Pages.Index.BackLabel)]"
                    Page="/" />
     </Toolbar>
     <PageContent>
@@ -52,7 +52,7 @@
                                     {
                                         <MudProgressCircular Class="mr-2" Color="Color.Primary" Indeterminate="true" Size="Size.Small" />
                                     }
-                                    <MudText>@Localizer[nameof(Resources.Pages.Tools.Index.StopBotLabel)]</MudText>
+                                    <MudText>@Localizer[nameof(Resources.Pages.Index.StopBotLabel)]</MudText>
                                 </MudButton>
                             }
                             else
@@ -62,7 +62,7 @@
                                     {
                                         <MudProgressCircular Class="mr-2" Color="Color.Primary" Indeterminate="true" Size="Size.Small" />
                                     }
-                                    <MudText>@Localizer[nameof(Resources.Pages.Tools.Index.StartBotLabel)]</MudText>
+                                    <MudText>@Localizer[nameof(Resources.Pages.Index.StartBotLabel)]</MudText>
                                 </MudButton>
                             }
                         </MudItem>
