+++ src/KITT.Web.App/Clients/ClientsServiceCollectionExtensions.cs
@@ -22,19 +22,4 @@ public static class ClientsServiceCollectionExtensions
 
         return services;
     }
-
-    public static IServiceCollection AddToolsClients(this IServiceCollection services, string clientBaseAddress)
-    {
-        services
-            .AddHttpClient<IBotClient, BotHttpClient>(client => client.BaseAddress = new Uri(clientBaseAddress))
-            .AddHttpMessageHandler<BaseAddressAuthorizationMessageHandler>();
-
-        services
-            .AddHttpClient<Tools.IStreamingsClient, Tools.Http.StreamingsHttpClient>(
-                "Tools.StreamingsClient", 
-                client => client.BaseAddress = new Uri(clientBaseAddress))
-            .AddHttpMessageHandler<BaseAddressAuthorizationMessageHandler>();
-
-        return services;
-    }
 }
