+++ src/KITT.Web.App/_Imports.razor
@@ -18,3 +18,5 @@
 @using Microsoft.Extensions.Localization
 @using MudBlazor
 @using KITT.Web.App.Components
+@using KITT.Web.App.UI.Layouts 
+@using KITT.Web.App.UI.Components
